refactor(runner): modulize runner.cc
runner.cc is too large, sepreate RunnerBuilder, RunnerContext, Runner and ClusterTask in
difference files
