+++ hybridse/src/vm/generator.cc
@@ -16,6 +16,10 @@
 
 #include "vm/generator.h"
 
+#include <utility>
+
+#include "node/sql_node.h"
+#include "vm/catalog.h"
 #include "vm/catalog_wrapper.h"
 #include "vm/runner.h"
 
@@ -233,10 +237,12 @@ Row JoinGenerator::RowLastJoinDropLeftSlices(
     return right_row;
 }
 
-std::shared_ptr<PartitionHandler> JoinGenerator::LazyLastJoin(std::shared_ptr<PartitionHandler> left,
-                                                              std::shared_ptr<PartitionHandler> right,
-                                                              const Row& parameter) {
-    return std::make_shared<LazyLastJoinPartitionHandler>(left, right, parameter, shared_from_this());
+
+
+std::shared_ptr<PartitionHandler> JoinGenerator::LazyJoinOptimized(std::shared_ptr<PartitionHandler> left,
+                                                                   std::shared_ptr<PartitionHandler> right,
+                                                                   const Row& parameter) {
+    return std::make_shared<LazyJoinPartitionHandler>(left, right, parameter, shared_from_this());
 }
 
 Row JoinGenerator::RowLastJoin(const Row& left_row,
@@ -276,6 +282,7 @@ Row JoinGenerator::RowLastJoinPartition(
     auto right_table = partition->GetSegment(partition_key);
     return RowLastJoinTable(left_row, right_table, parameter);
 }
+
 Row JoinGenerator::RowLastJoinTable(const Row& left_row,
                                     std::shared_ptr<TableHandler> table,
                                     const Row& parameter) {
@@ -730,6 +737,103 @@ std::shared_ptr<DataHandler> FilterGenerator::Filter(std::shared_ptr<TableHandle
 bool FilterGenerator::ValidIndex() const {
     return index_seek_gen_.Valid();
 }
+std::vector<std::shared_ptr<DataHandler>> InputsGenerator::RunInputs(
+    RunnerContext& ctx) {
+    std::vector<std::shared_ptr<DataHandler>> union_inputs;
+    for (auto runner : input_runners_) {
+        union_inputs.push_back(runner->RunWithCache(ctx));
+    }
+    return union_inputs;
+}
 
+std::vector<std::shared_ptr<PartitionHandler>> WindowUnionGenerator::PartitionEach(
+    std::vector<std::shared_ptr<DataHandler>> union_inputs, const Row& parameter) {
+    std::vector<std::shared_ptr<PartitionHandler>> union_partitions;
+    if (!windows_gen_.empty()) {
+        union_partitions.reserve(windows_gen_.size());
+        for (size_t i = 0; i < inputs_cnt_; i++) {
+            union_partitions.push_back(
+                windows_gen_[i].partition_gen_.Partition(union_inputs[i], parameter));
+        }
+    }
+    return union_partitions;
+}
+
+std::vector<std::shared_ptr<DataHandler>> WindowJoinGenerator::RunInputs(
+    RunnerContext& ctx) {
+    std::vector<std::shared_ptr<DataHandler>> union_inputs;
+    if (!input_runners_.empty()) {
+        for (auto runner : input_runners_) {
+            union_inputs.push_back(runner->RunWithCache(ctx));
+        }
+    }
+    return union_inputs;
+}
+Row WindowJoinGenerator::Join(
+    const Row& left_row,
+    const std::vector<std::shared_ptr<DataHandler>>& join_right_tables,
+    const Row& parameter) {
+    Row row = left_row;
+    for (size_t i = 0; i < join_right_tables.size(); i++) {
+        row = joins_gen_[i]->RowLastJoin(row, join_right_tables[i], parameter);
+    }
+    return row;
+}
+
+void WindowJoinGenerator::AddWindowJoin(const class Join& join, size_t left_slices, Runner* runner) {
+    size_t right_slices = runner->output_schemas()->GetSchemaSourceSize();
+    joins_gen_.push_back(JoinGenerator::Create(join, left_slices, right_slices));
+    AddInput(runner);
+}
+
+std::vector<std::shared_ptr<TableHandler>> RequestWindowUnionGenerator::GetRequestWindows(
+    const Row& row, const Row& parameter, std::vector<std::shared_ptr<DataHandler>> union_inputs) {
+    std::vector<std::shared_ptr<TableHandler>> union_segments(union_inputs.size());
+    for (size_t i = 0; i < union_inputs.size(); i++) {
+        union_segments[i] = windows_gen_[i].GetRequestWindow(row, parameter, union_inputs[i]);
+    }
+    return union_segments;
+}
+void RequestWindowUnionGenerator::AddWindowUnion(const RequestWindowOp& window_op, Runner* runner) {
+    windows_gen_.emplace_back(window_op);
+    AddInput(runner);
+}
+void WindowUnionGenerator::AddWindowUnion(const WindowOp& window_op, Runner* runner) {
+    windows_gen_.push_back(WindowGenerator(window_op));
+    AddInput(runner);
+}
+std::shared_ptr<TableHandler> RequestWindowGenertor::GetRequestWindow(const Row& row, const Row& parameter,
+                                                                      std::shared_ptr<DataHandler> input) {
+    auto segment = index_seek_gen_.SegmentOfKey(row, parameter, input);
+
+    if (filter_gen_.Valid()) {
+        auto filter_key = filter_gen_.GetKey(row, parameter);
+        segment = filter_gen_.Filter(parameter, segment, filter_key);
+    }
+    if (sort_gen_.Valid()) {
+        segment = sort_gen_.Sort(segment, true);
+    }
+    return segment;
+}
+std::shared_ptr<TableHandler> FilterKeyGenerator::Filter(const Row& parameter, std::shared_ptr<TableHandler> table,
+                                                         const std::string& request_keys) {
+    if (!filter_key_.Valid()) {
+        return table;
+    }
+    auto mem_table = std::shared_ptr<MemTimeTableHandler>(new MemTimeTableHandler());
+    mem_table->SetOrderType(table->GetOrderType());
+    auto iter = table->GetIterator();
+    if (iter) {
+        iter->SeekToFirst();
+        while (iter->Valid()) {
+            std::string keys = filter_key_.Gen(iter->GetValue(), parameter);
+            if (request_keys == keys) {
+                mem_table->AddRow(iter->GetKey(), iter->GetValue());
+            }
+            iter->Next();
+        }
+    }
+    return mem_table;
+}
 }  // namespace vm
 }  // namespace hybridse
